pyproject.toml
[project]
name = "diabeteshospital"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "annotated-types==0.7.0",
    "asttokens==3.0.0",
    "attrs==25.3.0",
    "bottleneck==1.6.0",
    "catboost==1.2.8",
    "certifi==2025.8.3",
    "charset-normalizer==3.4.3",
    "comm==0.2.3",
    "contourpy==1.3.3",
    "cycler==0.12.1",
    "dacite==1.9.2",
    "debugpy==1.8.16",
    "decorator==5.2.1",
    "executing==2.2.1",
    "fonttools==4.59.2",
    "graphviz==0.21",
    "htmlmin==0.1.12",
    "idna==3.10",
    "imagehash==4.3.1",
    "ipykernel==6.30.1",
    "ipython==9.5.0",
    "ipython-pygments-lexers==1.1.1",
    "jedi==0.19.2",
    "jinja2==3.1.6",
    "joblib==1.5.2",
    "jupyter-client==8.6.3",
    "jupyter-core==5.8.1",
    "jupyterlab>=4.4.7",
    "kiwisolver==1.4.9",
    "llvmlite==0.44.0",
    "markupsafe==3.0.2",
    "matplotlib==3.10.0",
    "matplotlib-inline==0.1.7",
    "multimethod==1.12",
    "narwhals==2.4.0",
    "nest-asyncio==1.6.0",
    "networkx==3.5",
    "numba==0.61.0",
    "numexpr==2.12.1",
    "numpy<2",
    "nvidia-nccl-cu12==2.28.3",
    "packaging==25.0",
    "pandas==2.3.2",
    "parso==0.8.5",
    "patsy==1.0.1",
    "pexpect==4.9.0",
    "phik==0.12.5",
    "pillow==11.3.0",
    "platformdirs==4.4.0",
    "plotly==6.3.0",
    "prompt-toolkit==3.0.52",
    "psutil==7.0.0",
    "ptyprocess==0.7.0",
    "pure-eval==0.2.3",
    "puremagic==1.30",
    "pybind11==3.0.1",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pygments==2.19.2",
    "pyparsing==3.2.3",
    "python-dateutil==2.9.0.post0",
    "pytz==2025.2",
    "pywavelets==1.9.0",
    "pyyaml==6.0.2",
    "pyzmq==27.1.0",
    "requests==2.32.5",
    "scikit-learn==1.7.2",
    "scipy==1.15.3",
    "seaborn==0.13.2",
    "six==1.17.0",
    "stack-data==0.6.3",
    "statsmodels==0.14.5",
    "threadpoolctl==3.6.0",
    "tornado==6.5.2",
    "tqdm==4.67.1",
    "traitlets==5.14.3",
    "typeguard==4.4.4",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.1",
    "tzdata==2025.2",
    "urllib3==2.5.0",
    "visions==0.8.1",
    "wcwidth==0.2.13",
    "wordcloud==1.9.4",
    "xgboost==3.0.5",
    "ydata-profiling==4.16.1",
]
